requirements.txt
requests>=2.32.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
tensorflow-cpu>=2.17.0
fastapi>=0.115.0
uvicorn>=0.30.0
tqdm>=4.66.0

